Ventanas Pedidos y Lógica

+++ Ferreteria/Entities/PedidoClienteE.cs
@@ -21,7 +21,7 @@ namespace Entities
             LinkedList<LinkedList<PedidoClienteServicioE>> servicios,
             DateTime horaRecibidoBodega, DateTime horaSalidaBodega, string codigoConductor,
             DateTime horaIniciaTransporte, DateTime horaFinalizaTransporte, DateTime fechaFactura,
-            string codigoCajero, double subTotal, double iva, double total, string estado)
+            string codigoCajero, decimal subTotal, decimal iva, decimal total, string estado)
             : base(id, cedulaCliente, horaRecibidoBodega, horaSalidaBodega, codigoConductor,
                   horaIniciaTransporte, horaFinalizaTransporte, fechaFactura, codigoCajero, subTotal, iva, total)
         {
